NewVersion/requirements.txt
asgiref
django
et-xmlfile
mysqlclient
openpyxl
sqlparse
tzdata
pandas
django-import-export
uwsgi
django-stubs
